Split FileBrowser and OverwritePrompt from LoadSave unit
